requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
aiofiles
a2wsgi
